requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
Django==4.2.7
django-environ==0.11.2
et-xmlfile==1.1.0
filelock==3.13.4
fsspec==2024.3.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
install==1.3.5
jaconv==0.3.4
Janome==0.5.0
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
mecab-python3==1.0.9
mojimoji==0.0.13
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.20.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pillow==11.0.0
platformdirs==4.2.1
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pyparsing==3.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.4.16
requests==2.31.0
romaji==0.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.3
sqlparse==0.4.4
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.0
wrapt==1.16.0
